cleanup, mostly renaming *Impl classes
