+++ common/src/main/java/org/opentestsystem/rdw/ingest/common/repository/JdbcImportRepository.java
@@ -1,7 +1,6 @@
-package org.opentestsystem.rdw.ingest.common.repository.impl;
+package org.opentestsystem.rdw.ingest.common.repository;
 
 import org.opentestsystem.rdw.ingest.common.model.ImportStatus;
-import org.opentestsystem.rdw.ingest.common.repository.ImportRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -9,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-class ImportRepositoryImpl implements ImportRepository {
+class JdbcImportRepository implements ImportRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
@@ -17,16 +16,18 @@ class ImportRepositoryImpl implements ImportRepository {
     private String updateStatusByIdSql;
 
     @Autowired
-    public ImportRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    public JdbcImportRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    public void updateStatusAndMessageById(long id, final ImportStatus importStatus, final String message) {
-        int updatedCount = jdbcTemplate.update(updateStatusByIdSql, new MapSqlParameterSource()
+    public void updateStatusAndMessageById(final long id, final ImportStatus importStatus, final String message) {
+        final int updatedCount = jdbcTemplate.update(updateStatusByIdSql, new MapSqlParameterSource()
                 .addValue("id", id)
                 .addValue("status", importStatus.getValue())
                 .addValue("message", message));
-        if (updatedCount == 0) throw new IllegalArgumentException("unknown import id [" + id + "]");
+        if (updatedCount == 0) {
+            throw new IllegalArgumentException("unknown import id [" + id + "]");
+        }
     }
 }
