+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcAdminConditionRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
 
-import org.opentestsystem.rdw.ingest.processor.repository.GenderRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.AdminConditionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -11,25 +11,25 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class GenderRepositoryImpl implements GenderRepository {
+class JdbcAdminConditionRepository implements AdminConditionRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.gender.findIdByName}")
+    @Value("${sql.administrationCondition.findIdByName}")
     private String findIdByNameSql;
 
     @Autowired
-    GenderRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcAdminConditionRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    @Cacheable(value = "gender")
+    @Cacheable(value = "adminCondition")
     public Integer findIdByName(final String name) {
         try {
             return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown gender name [" + name + "]");
+            throw new IllegalArgumentException("unknown administration condition name [" + name + "]");
         }
     }
 }
