+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcAssessmentRepository.java
@@ -18,7 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 import static com.google.common.collect.Maps.newHashMap;
 
 @Repository
-class AssessmentRepositoryImpl implements AssessmentRepository {
+class JdbcAssessmentRepository implements AssessmentRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
@@ -32,7 +32,7 @@ class AssessmentRepositoryImpl implements AssessmentRepository {
     private String sqlfindAllItemsForAssessment;
 
     @Autowired
-    AssessmentRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcAssessmentRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
