+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcAssessmentTypeRepository.java
@@ -1,7 +1,6 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
-
-import org.opentestsystem.rdw.ingest.processor.repository.AdminConditionRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.AssessmentTypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -11,25 +10,25 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class AdminConditionRepositoryImpl implements AdminConditionRepository {
+class JdbcAssessmentTypeRepository implements AssessmentTypeRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.administrationCondition.findIdByName}")
-    private String findIdByNameSql;
+    @Value("${sql.asmtType.findCodeById}")
+    private String findCodeByIdSql;
 
     @Autowired
-    AdminConditionRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcAssessmentTypeRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    @Cacheable(value = "adminCondition")
-    public Integer findIdByName(final String name) {
+    @Cacheable(value = "assessmentType")
+    public String findCodeById(final int id) {
         try {
-            return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
+            return jdbcTemplate.queryForObject(findCodeByIdSql, new MapSqlParameterSource("id", id), String.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown administration condition name [" + name + "]");
+            throw new IllegalArgumentException("unknown assessment type id [" + id + "]");
         }
     }
 }
