+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcCompletenessRepository.java
@@ -1,8 +1,7 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
 
-import java.util.List;
-import org.opentestsystem.rdw.ingest.processor.repository.EthnicityRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.CompletenessRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -12,33 +11,25 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class EthnicityRepositoryImpl implements EthnicityRepository {
+class JdbcCompletenessRepository implements CompletenessRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.ethnicity.findIdByName}")
+    @Value("${sql.completeness.findIdByName}")
     private String findIdByNameSql;
 
-    @Value("${sql.ethnicity.findAllNames}")
-    private String findAllNamesSql;
-
     @Autowired
-    EthnicityRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcCompletenessRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    @Cacheable(value = "ethnicity")
+    @Cacheable(value = "completeness")
     public Integer findIdByName(final String name) {
         try {
             return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown ethnicity name [" + name + "]");
+            throw new IllegalArgumentException("unknown completeness name [" + name + "]");
         }
-    }
-
-    @Override
-    public List<String> findAllNames() {
-        return jdbcTemplate.queryForList(findAllNamesSql, new MapSqlParameterSource(), String.class);
     }
 }
