+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcEthnicityRepository.java
@@ -1,7 +1,8 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
 
-import org.opentestsystem.rdw.ingest.processor.repository.CompletenessRepository;
+import java.util.List;
+import org.opentestsystem.rdw.ingest.processor.repository.EthnicityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -11,25 +12,33 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class CompletnessRepositoryImpl implements CompletenessRepository {
+class JdbcEthnicityRepository implements EthnicityRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.completeness.findIdByName}")
+    @Value("${sql.ethnicity.findIdByName}")
     private String findIdByNameSql;
 
+    @Value("${sql.ethnicity.findAllNames}")
+    private String findAllNamesSql;
+
     @Autowired
-    CompletnessRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcEthnicityRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    @Cacheable(value = "completeness")
+    @Cacheable(value = "ethnicity")
     public Integer findIdByName(final String name) {
         try {
             return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown completeness name [" + name + "]");
+            throw new IllegalArgumentException("unknown ethnicity name [" + name + "]");
         }
+    }
+
+    @Override
+    public List<String> findAllNames() {
+        return jdbcTemplate.queryForList(findAllNamesSql, new MapSqlParameterSource(), String.class);
     }
 }
