+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcExamRepository.java
@@ -1,10 +1,11 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
+import org.opentestsystem.rdw.ingest.processor.model.Exam;
+import org.opentestsystem.rdw.ingest.processor.model.ExamClaim;
 import org.opentestsystem.rdw.ingest.processor.model.ExamItem;
 import org.opentestsystem.rdw.ingest.processor.model.ExamItemSubScore;
-import org.opentestsystem.rdw.ingest.processor.model.IabExam;
 import org.opentestsystem.rdw.ingest.processor.model.StudentExamAttributes;
-import org.opentestsystem.rdw.ingest.processor.repository.IabExamRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.ExamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -18,34 +19,36 @@ import org.springframework.transaction.annotation.Transactional;
 import java.sql.Timestamp;
 import java.util.List;
 
-
 @Repository
-public class IabExamRepositoryImpl implements IabExamRepository {
+class JdbcExamRepository implements ExamRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.iabExam.iabExamStudent.create}")
+    @Value("${sql.exam.examClaimScore.create}")
+    private String sqlExamClaimScoreCreate;
+
+    @Value("${sql.exam.examStudent.create}")
     private String sqlExamStudentCreate;
 
-    @Value("${sql.iabExam.iabExamItem.create}")
+    @Value("${sql.exam.examItem.create}")
     private String sqlExamItemCreate;
 
-    @Value("${sql.iabExam.create}")
+    @Value("${sql.exam.create}")
     private String sqlCreate;
 
     @Autowired
-    IabExamRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcExamRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Transactional
     @Override
-    public long create(final IabExam exam) {
+    public long create(final Exam exam) {
         final long studentExamId = create(exam.getStudentExamAttributes());
 
         final KeyHolder keyHolder = new GeneratedKeyHolder();
         final SqlParameterSource parameterSource = new MapSqlParameterSource()
-                .addValue("iab_exam_student_id", studentExamId)
+                .addValue("exam_student_id", studentExamId)
                 .addValue("school_year", exam.getSchoolYear())
                 .addValue("import_id", exam.getImportId())
                 .addValue("asmt_id", exam.getAssessmentId())
@@ -57,13 +60,28 @@ public class IabExamRepositoryImpl implements IabExamRepository {
                 .addValue("session_id", exam.getSessionId())
                 .addValue("scale_score", exam.getScaleScore())
                 .addValue("scale_score_std_err", exam.getScaleScoreStdErr())
-                .addValue("category", exam.getCategory())
+                .addValue("achievement_level", exam.getAchievementLevel())
                 .addValue("completed_at", Timestamp.from(exam.getCompletedAt()));
 
         jdbcTemplate.update(sqlCreate, parameterSource, keyHolder);
         final Long id = keyHolder.getKey().longValue();
+        batchCreateExamClaims(exam.getExamClaims(), id);
         batchCreateExamItems(exam.getExamItems(), id);
         return id;
+    }
+
+    private void batchCreateExamClaims(final List<ExamClaim> examClaims, final long examId) {
+        if (examClaims.isEmpty()) return;
+
+        final SqlParameterSource[] batchParameters = examClaims.stream()
+                .map(examClaim -> new MapSqlParameterSource("exam_id", examId)
+                        .addValue("subject_claim_score_id", examClaim.getClaimId())
+                        .addValue("scale_score", examClaim.getScaleScore())
+                        .addValue("scale_score_std_err", examClaim.getScaleScoreStdErr())
+                        .addValue("category", examClaim.getCategory()))
+                .toArray(MapSqlParameterSource[]::new);
+
+        jdbcTemplate.batchUpdate(sqlExamClaimScoreCreate, batchParameters);
     }
 
     private long create(final StudentExamAttributes studentExamAttributes) {
@@ -92,7 +110,7 @@ public class IabExamRepositoryImpl implements IabExamRepository {
         final MapSqlParameterSource[] batchParameters = new MapSqlParameterSource[examItems.size()];
         for (int i = 0; i < examItems.size(); i++) {
             final ExamItem examItem = examItems.get(i);
-            batchParameters[i] = new MapSqlParameterSource("iab_exam_id", examId)
+            batchParameters[i] = new MapSqlParameterSource("exam_id", examId)
                     .addValue("item_id", examItem.getItemId())
                     .addValue("response", examItem.getResponse())
                     .addValue("score_status", examItem.getScoreStatus())
