+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcGenderRepository.java
@@ -1,6 +1,7 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
-import org.opentestsystem.rdw.ingest.processor.repository.AssessmentTypeRepository;
+
+import org.opentestsystem.rdw.ingest.processor.repository.GenderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -10,25 +11,25 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-class AssessmentTypeRepositoryImpl implements AssessmentTypeRepository {
+class JdbcGenderRepository implements GenderRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.asmtType.findCodeById}")
-    private String findCodeByIdSql;
+    @Value("${sql.gender.findIdByName}")
+    private String findIdByNameSql;
 
     @Autowired
-    AssessmentTypeRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcGenderRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    @Cacheable(value = "assessmentType")
-    public String findCodeById(final int id) {
+    @Cacheable(value = "gender")
+    public Integer findIdByName(final String name) {
         try {
-            return jdbcTemplate.queryForObject(findCodeByIdSql, new MapSqlParameterSource("id", id), String.class);
+            return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown assessment type id [" + id + "]");
+            throw new IllegalArgumentException("unknown gender name [" + name + "]");
         }
     }
 }
