+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcIabExamRepository.java
@@ -1,11 +1,10 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
-import org.opentestsystem.rdw.ingest.processor.model.Exam;
-import org.opentestsystem.rdw.ingest.processor.model.ExamClaim;
 import org.opentestsystem.rdw.ingest.processor.model.ExamItem;
 import org.opentestsystem.rdw.ingest.processor.model.ExamItemSubScore;
+import org.opentestsystem.rdw.ingest.processor.model.IabExam;
 import org.opentestsystem.rdw.ingest.processor.model.StudentExamAttributes;
-import org.opentestsystem.rdw.ingest.processor.repository.ExamRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.IabExamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -19,36 +18,34 @@ import org.springframework.transaction.annotation.Transactional;
 import java.sql.Timestamp;
 import java.util.List;
 
+
 @Repository
-public class ExamRepositoryImpl implements ExamRepository {
+class JdbcIabExamRepository implements IabExamRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.exam.examClaimScore.create}")
-    private String sqlExamClaimScoreCreate;
-
-    @Value("${sql.exam.examStudent.create}")
+    @Value("${sql.iabExam.iabExamStudent.create}")
     private String sqlExamStudentCreate;
 
-    @Value("${sql.exam.examItem.create}")
+    @Value("${sql.iabExam.iabExamItem.create}")
     private String sqlExamItemCreate;
 
-    @Value("${sql.exam.create}")
+    @Value("${sql.iabExam.create}")
     private String sqlCreate;
 
     @Autowired
-    ExamRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcIabExamRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Transactional
     @Override
-    public long create(final Exam exam) {
+    public long create(final IabExam exam) {
         final long studentExamId = create(exam.getStudentExamAttributes());
 
         final KeyHolder keyHolder = new GeneratedKeyHolder();
         final SqlParameterSource parameterSource = new MapSqlParameterSource()
-                .addValue("exam_student_id", studentExamId)
+                .addValue("iab_exam_student_id", studentExamId)
                 .addValue("school_year", exam.getSchoolYear())
                 .addValue("import_id", exam.getImportId())
                 .addValue("asmt_id", exam.getAssessmentId())
@@ -60,28 +57,13 @@ public class ExamRepositoryImpl implements ExamRepository {
                 .addValue("session_id", exam.getSessionId())
                 .addValue("scale_score", exam.getScaleScore())
                 .addValue("scale_score_std_err", exam.getScaleScoreStdErr())
-                .addValue("achievement_level", exam.getAchievementLevel())
+                .addValue("category", exam.getCategory())
                 .addValue("completed_at", Timestamp.from(exam.getCompletedAt()));
 
         jdbcTemplate.update(sqlCreate, parameterSource, keyHolder);
         final Long id = keyHolder.getKey().longValue();
-        batchCreateExamClaims(exam.getExamClaims(), id);
         batchCreateExamItems(exam.getExamItems(), id);
         return id;
-    }
-
-    private void batchCreateExamClaims(final List<ExamClaim> examClaims, final long examId) {
-        if (examClaims.isEmpty()) return;
-
-        final SqlParameterSource[] batchParameters = examClaims.stream()
-                .map(examClaim -> new MapSqlParameterSource("exam_id", examId)
-                        .addValue("subject_claim_score_id", examClaim.getClaimId())
-                        .addValue("scale_score", examClaim.getScaleScore())
-                        .addValue("scale_score_std_err", examClaim.getScaleScoreStdErr())
-                        .addValue("category", examClaim.getCategory()))
-                .toArray(MapSqlParameterSource[]::new);
-
-        jdbcTemplate.batchUpdate(sqlExamClaimScoreCreate, batchParameters);
     }
 
     private long create(final StudentExamAttributes studentExamAttributes) {
@@ -110,7 +92,7 @@ public class ExamRepositoryImpl implements ExamRepository {
         final MapSqlParameterSource[] batchParameters = new MapSqlParameterSource[examItems.size()];
         for (int i = 0; i < examItems.size(); i++) {
             final ExamItem examItem = examItems.get(i);
-            batchParameters[i] = new MapSqlParameterSource("exam_id", examId)
+            batchParameters[i] = new MapSqlParameterSource("iab_exam_id", examId)
                     .addValue("item_id", examItem.getItemId())
                     .addValue("response", examItem.getResponse())
                     .addValue("score_status", examItem.getScoreStatus())
