+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcItemTraitRepository.java
@@ -13,14 +13,14 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class ItemTraitRepositoryImpl implements ItemTraitRepository {
+class JdbcItemTraitRepository implements ItemTraitRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.itemTraitScore.findAll}")
     private String findAllSql;
 
     @Autowired
-    ItemTraitRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcItemTraitRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
