+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcSchoolRepository.java
@@ -10,12 +10,12 @@ import org.springframework.jdbc.core.simple.SimpleJdbcCall;
 import org.springframework.stereotype.Repository;
 
 @Repository
-class SchoolRepositoryImpl implements SchoolRepository {
+class JdbcSchoolRepository implements SchoolRepository {
 
     private final JdbcTemplate jdbcTemplate;
 
     @Autowired
-    SchoolRepositoryImpl(final JdbcTemplate jdbcTemplate) {
+    JdbcSchoolRepository(final JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
