+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcStubbedOutAssessmentRepository.java
@@ -16,10 +16,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 
 @Repository
-class StubbedOutAssessmentRepositoryImpl implements StubbedOutAssessmentRepository {
+class JdbcStubbedOutAssessmentRepository implements StubbedOutAssessmentRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
-    private AssessmentRepository assessmentRepository;
+    private final AssessmentRepository assessmentRepository;
 
     @Value("${sql.assessment.create}")
     private String sqlCreate;
@@ -28,7 +28,7 @@ class StubbedOutAssessmentRepositoryImpl implements StubbedOutAssessmentReposito
     private String sqlCreateItems;
 
     @Autowired
-    StubbedOutAssessmentRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate,
+    JdbcStubbedOutAssessmentRepository(final NamedParameterJdbcTemplate jdbcTemplate,
                                        final AssessmentRepository assessmentRepository) {
         this.jdbcTemplate = jdbcTemplate;
         this.assessmentRepository = assessmentRepository;
