+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcStudentGroupRepository.java
@@ -24,7 +24,7 @@ import java.util.List;
  * JDBC implementation of {@link StudentGroupRepository}.
  */
 @Repository
-public class StudentGroupRepositoryImpl implements StudentGroupRepository {
+class JdbcStudentGroupRepository implements StudentGroupRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final StudentGroupRowMapper mapper = new StudentGroupRowMapper();
@@ -39,7 +39,7 @@ public class StudentGroupRepositoryImpl implements StudentGroupRepository {
     private String sqlAddStudentMembership;
 
     @Autowired
-    StudentGroupRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcStudentGroupRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
