+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcStudentRepository.java
@@ -16,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import static com.google.common.collect.Lists.newArrayList;
 
 @Repository
-class StudentRepositoryImpl implements StudentRepository {
+class JdbcStudentRepository implements StudentRepository {
 
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
     private final JdbcTemplate jdbcTemplate;
@@ -34,7 +34,7 @@ class StudentRepositoryImpl implements StudentRepository {
     private String updateStudentImportIdSql;
 
     @Autowired
-    StudentRepositoryImpl(final NamedParameterJdbcTemplate namedParameterJdbcTemplate, final JdbcTemplate jdbcTemplate) {
+    JdbcStudentRepository(final NamedParameterJdbcTemplate namedParameterJdbcTemplate, final JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
         this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
     }
@@ -70,7 +70,7 @@ class StudentRepositoryImpl implements StudentRepository {
         namedParameterJdbcTemplate.update(deleteStudentEthnicitySql, new MapSqlParameterSource("student_id", studentId));
         if (!newIds.isEmpty()) {
             final List<SqlParameterSource> batchParameters = newArrayList();
-            for (int newId : newIds) {
+            for (final int newId : newIds) {
                 batchParameters.add(new MapSqlParameterSource("student_id", studentId).addValue("ethnicity_id", newId));
             }
             namedParameterJdbcTemplate.batchUpdate(createStudentEthnicitySql, batchParameters.toArray(new SqlParameterSource[batchParameters.size()]));
