+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcSubjectRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
 
-import org.opentestsystem.rdw.ingest.processor.repository.GradeRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.SubjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -11,24 +11,24 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class GradeRepositoryImpl implements GradeRepository {
+class JdbcSubjectRepository implements SubjectRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.grade.findIdByCode}")
-    private String findIdByCodeSql;
+    @Value("${sql.subject.findIdByName}")
+    private String findIdByNameSql;
 
     @Autowired
-    GradeRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcSubjectRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
     @Override
-    @Cacheable(value = "grade")
-    public Integer findIdByCode(final String code) {
+    @Cacheable(value = "subject")
+    public Integer findIdByName(final String name) {
         try {
-            return jdbcTemplate.queryForObject(findIdByCodeSql, new MapSqlParameterSource("code", code), Integer.class);
+            return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown grade code [" + code + "]");
+            throw new IllegalArgumentException("unknown subject name [" + name + "]");
         }
     }
 }
