+++ exam-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/repository/impl/AssessmentRepositoryIT.java
@@ -19,7 +19,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest
 @Transactional
 @ActiveProfiles("test")
-public class AssessmentRepositoryImplIT {
+public class AssessmentRepositoryIT {
 
     @Autowired
     private AssessmentRepository repository;
