+++ exam-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/repository/impl/StudentGroupRepositoryIT.java
@@ -31,7 +31,7 @@ import static org.springframework.test.jdbc.JdbcTestUtils.countRowsInTable;
                 "(-100, '100', 'Smith', 'Alice', 2, '2004-09-02', -99), " +
                 "(-101, '101', 'Jones', 'Bob', 1, '2004-03-02', -99)"
 })
-public class StudentGroupRepositoryImplIT {
+public class StudentGroupRepositoryIT {
 
     @Autowired
     private StudentGroupRepository repository;
