+++ import-service/src/main/java/org/opentestsystem/rdw/ingest/repository/JdbcRdwImportRepository.java
@@ -23,7 +23,7 @@ import java.util.List;
 import static com.google.common.collect.Lists.newArrayList;
 
 /**
- * Default implementation of {@link RdwImportRepository}.
+ * JDBC implementation of {@link RdwImportRepository}.
  */
 @Repository
 class JdbcRdwImportRepository implements RdwImportRepository {
