+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/AssessmentTypeRepository.java
@@ -9,7 +9,6 @@ public interface AssessmentTypeRepository {
      *
      * @param code the assessment type code, e.g. "ica"
      * @return the id if found, or null if not found.
-     * @throws IllegalArgumentException
      */
     Integer findIdByCode(String code);
 }
