+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/DepthOfKnowledgeRepository.java
@@ -10,8 +10,7 @@ public interface DepthOfKnowledgeRepository {
      *
      * @param level the dok level
      * @param subjectName the corresponding subject name of the for this item
-     * @return the id if it exists
-     * @throws IllegalArgumentException
+     * @return the id if it exists, null if it doesn't
      */
     Integer findIdByLevelAndSubject(int level, String subjectName);
 }
