+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/GradeRepository.java
@@ -8,8 +8,7 @@ public interface GradeRepository {
      * Finds an id of the grade by its code
      *
      * @param code the grade code, e.g. KG, 01, 02, etc.
-     * @return the id
-     * @throws IllegalArgumentException
+     * @return the id, null if it doesn't exist
      */
     Integer findIdByCode(String code);
 }
