+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/ItemRepository.java
@@ -10,8 +10,7 @@ public interface ItemRepository {
      * Finds an id of the item by its item_key
      *
      * @param item_key the item of the subject
-     * @return the id
-     * @throws IllegalArgumentException if the given name is not found
+     * @return the id, null if it doesn't exist
      */
     Integer findIdByKey(String item_key);
