+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/MathPracticeRepository.java
@@ -8,8 +8,7 @@ public interface MathPracticeRepository {
      * Finds an id of the target by its code
      *
      * @param practice the practice of the target
-     * @return the practice if it exists
-     * @throws IllegalArgumentException
+     * @return the practice if it exists, null if it doesn't
      */
     Integer findByPractice(int practice);
 }
