+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/SubjectRepository.java
@@ -8,8 +8,7 @@ public interface SubjectRepository {
      * Finds an id of the subject by its name
      *
      * @param name the name of the subject
-     * @return the id
-     * @throws IllegalArgumentException if the given name is not found
+     * @return the id if it exists, null if it doesn't
      */
     Integer findIdByName(String name);
 }
