+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/TargetRepository.java
@@ -9,8 +9,7 @@ public interface TargetRepository {
      * Finds an id of the target by its code
      *
      * @param code the code of the target
-     * @return the id
-     * @throws IllegalArgumentException
+     * @return the id if it exists, null if it doesn't
      */
     Integer findIdByCodeAndClaim(String code, String claimId);
 }
