+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcAssessmentPackageRepository.java
@@ -10,12 +10,8 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-/**
- * AssessmentPackageRepositoryImpl - Provides for bulk inserting an assessment "package", which is a
- * collection of Assessments and corresponding Items
- */
 @Repository
-public class AssessmentPackageRepositoryImpl implements AssessmentPackageRepository {
+class JdbcAssessmentPackageRepository implements AssessmentPackageRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
@@ -29,7 +25,7 @@ public class AssessmentPackageRepositoryImpl implements AssessmentPackageReposit
     private String sqlAssessmentCreate;
 
     @Autowired
-    public AssessmentPackageRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    public JdbcAssessmentPackageRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
