+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcAssessmentRepository.java
@@ -6,15 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * AssessmentRepositoryImpl - provides interaction with the assessment table
- */
 @Repository
-public class AssessmentRepositoryImpl implements AssessmentRepository {
+class JdbcAssessmentRepository implements AssessmentRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Autowired
-    AssessmentRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcAssessmentRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -28,6 +25,4 @@ public class AssessmentRepositoryImpl implements AssessmentRepository {
     public Assessment findOneByNaturalId(String id) {
         throw new UnsupportedOperationException();
     }
-
-
 }
