+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcAssessmentTypeRepository.java
@@ -9,35 +9,30 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * GradeRepositoryImpl - provides interaction with the grade table
- */
 @Repository
-public class AssessmentTypeRepositoryImpl implements AssessmentTypeRepository {
+class JdbcAssessmentTypeRepository implements AssessmentTypeRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.asmtType.findIdByCode}")
     private String findIdByCodeSql;
 
     @Autowired
-    AssessmentTypeRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcAssessmentTypeRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
-
 
     /**
      * Finds an id of the type by its code
      *
      * @param code the assessment type code, e.g. "ica"
      * @return the id if found, or null if not found.
-     * @throws IllegalArgumentException
      */
     @Override
     @Cacheable(value = "asmtType")
     public Integer findIdByCode(final String code) {
         try {
             return jdbcTemplate.queryForObject(findIdByCodeSql, new MapSqlParameterSource("code", code), Integer.class);
-        } catch (EmptyResultDataAccessException ignore) {
+        } catch (final EmptyResultDataAccessException ignore) {
             return null;
         }
     }
