+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcClaimRepository.java
@@ -9,18 +9,15 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * ClaimRepositoryImpl - provides interaction with the claims table
- */
 @Repository
-public class ClaimRepositoryImpl implements ClaimRepository {
+class JdbcClaimRepository implements ClaimRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.claim.findIdByCode}")
     private String findIdByCodeSql;
 
     @Autowired
-    public ClaimRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    public JdbcClaimRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -36,7 +33,7 @@ public class ClaimRepositoryImpl implements ClaimRepository {
     public Integer findIdByCode(final String code) {
         try {
             return jdbcTemplate.queryForObject(findIdByCodeSql, new MapSqlParameterSource("code", code), Integer.class);
-        } catch (EmptyResultDataAccessException ignore) {
+        } catch (final EmptyResultDataAccessException ignore) {
             return null;
         }
     }
