+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcDepthOfKnowledgeRepository.java
@@ -9,18 +9,15 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * DepthOfKnowledgeRepositoryImpl - provides interaction with the depth of knowledge table
- */
 @Repository
-public class DepthOfKnowledgeRepositoryImpl implements DepthOfKnowledgeRepository {
+class JdbcDepthOfKnowledgeRepository implements DepthOfKnowledgeRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.dok.findIdByLevelAndSubject}")
     private String findIdByLevelAndSubject;
 
     @Autowired
-    public DepthOfKnowledgeRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    public JdbcDepthOfKnowledgeRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -29,17 +26,16 @@ public class DepthOfKnowledgeRepositoryImpl implements DepthOfKnowledgeRepositor
      *
      * @param level the dok level
      * @param subjectName the corresponding subject name of the for this item
-     * @throws IllegalArgumentException
      */
     @Override
     @Cacheable(value = "dokId")
-    public Integer findIdByLevelAndSubject(int level, String subjectName) {
+    public Integer findIdByLevelAndSubject(final int level, final String subjectName) {
         try {
             return jdbcTemplate.queryForObject(findIdByLevelAndSubject,
                     new MapSqlParameterSource()
                             .addValue("level", level)
                             .addValue("subjectName", subjectName), Integer.class);
-        } catch (EmptyResultDataAccessException ignore) {
+        } catch (final EmptyResultDataAccessException ignore) {
             return null;
         }
     }
