+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcGradeRepository.java
@@ -1,7 +1,6 @@
 package org.opentestsystem.rdw.ingest.processor.repository.impl;
 
-
-import org.opentestsystem.rdw.ingest.processor.repository.SubjectRepository;
+import org.opentestsystem.rdw.ingest.processor.repository.GradeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
@@ -11,24 +10,30 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class SubjectRepositoryImpl implements SubjectRepository {
+class JdbcGradeRepository implements GradeRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
-    @Value("${sql.subject.findIdByName}")
-    private String findIdByNameSql;
+    @Value("${sql.grade.findIdByCode}")
+    private String findIdByCodeSql;
 
     @Autowired
-    SubjectRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcGradeRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
+    /**
+     * Finds an id of the grade by its code
+     *
+     * @param code the grade code, e.g. KG, 01, 02, etc.
+     * @return the id, or null if it doesn't exist
+     */
     @Override
-    @Cacheable(value = "subject")
-    public Integer findIdByName(final String name) {
+    @Cacheable(value = "grade")
+    public Integer findIdByCode(final String code) {
         try {
-            return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
+            return jdbcTemplate.queryForObject(findIdByCodeSql, new MapSqlParameterSource("code", code), Integer.class);
         } catch (final EmptyResultDataAccessException ignore) {
-            throw new IllegalArgumentException("unknown subject name [" + name + "]");
+            return null;
         }
     }
 }
