+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcItemRepository.java
@@ -7,16 +7,13 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-/**
- * ItemRepositoryImpl - provides interaction with the item table
- */
 @Repository
-public class ItemRepositoryImpl implements ItemRepository {
+class JdbcItemRepository implements ItemRepository {
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Autowired
-    public ItemRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    public JdbcItemRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -24,17 +21,16 @@ public class ItemRepositoryImpl implements ItemRepository {
      * Finds an id of the item by its item_key
      *
      * @param item_key the item of the subject
-     * @return the id
-     * @throws IllegalArgumentException if the given name is not found
+     * @return the id, or null if it doesn't exist
      */
     @Override
-    public Integer findIdByKey(String item_key) {
+    public Integer findIdByKey(final String item_key) {
         throw new UnsupportedOperationException();
     }
 
     @Override
     @Transactional
-    public void create(Item... items) {
+    public void create(final Item... items) {
         throw new UnsupportedOperationException();
     }
 }
