+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcMathPracticeRepository.java
@@ -9,18 +9,15 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * MathPracticeRepositoryImpl - provides interaction with the math practice table
- */
 @Repository
-public class MathPracticeRepositoryImpl implements MathPracticeRepository {
+class JdbcMathPracticeRepository implements MathPracticeRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.mathPractice.findByPractice}")
     private String findByPracticeSql;
 
     @Autowired
-    public MathPracticeRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    public JdbcMathPracticeRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -28,15 +25,14 @@ public class MathPracticeRepositoryImpl implements MathPracticeRepository {
      * Finds an practice of the target by its code
      *
      * @param practice the practice of the math practice
-     * @return the practice if it exists
-     * @throws IllegalArgumentException if the given practice is not found
+     * @return the practice if it exists, null if it doesn't
      */
     @Override
     @Cacheable(value = "mathPractice")
-    public Integer findByPractice(int practice) {
+    public Integer findByPractice(final int practice) {
         try {
             return jdbcTemplate.queryForObject(findByPracticeSql, new MapSqlParameterSource("practice", practice), Integer.class);
-        } catch (EmptyResultDataAccessException ignore) {
+        } catch (final EmptyResultDataAccessException ignore) {
             return null;
         }
     }
