+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcSubjectRepository.java
@@ -9,18 +9,15 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * SubjectRepositoryImpl - provides interaction with the subject table
- */
 @Repository
-public class SubjectRepositoryImpl implements SubjectRepository {
+class JdbcSubjectRepository implements SubjectRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.subject.findIdByName}")
     private String findIdByNameSql;
 
     @Autowired
-    SubjectRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcSubjectRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -36,7 +33,7 @@ public class SubjectRepositoryImpl implements SubjectRepository {
     public Integer findIdByName(final String name) {
         try {
             return jdbcTemplate.queryForObject(findIdByNameSql, new MapSqlParameterSource("name", name), Integer.class);
-        } catch (EmptyResultDataAccessException ignore) {
+        } catch (final EmptyResultDataAccessException ignore) {
             return null;
         }
     }
