+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcTargetRepository.java
@@ -9,18 +9,15 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-/**
- * TargetRepositoryImpl - provides interaction with the target table
- */
 @Repository
-public class TargetRepositoryImpl implements TargetRepository {
+class JdbcTargetRepository implements TargetRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Value("${sql.target.findIdByCodeAndClaim}")
     private String findIdByCodeAndClaimSql;
 
     @Autowired
-    TargetRepositoryImpl(final NamedParameterJdbcTemplate jdbcTemplate) {
+    JdbcTargetRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -28,19 +25,18 @@ public class TargetRepositoryImpl implements TargetRepository {
      * Finds an id of the target by its code
      *
      * @param code the code of the target
-     * @return the id
-     * @throws IllegalArgumentException if the given code is not found
+     * @return the id, null if it doesn't exist
      */
     @Override
     @Cacheable(value = "target")
-    public Integer findIdByCodeAndClaim(String code, String claimCode) {
+    public Integer findIdByCodeAndClaim(final String code, final String claimCode) {
         //ClaimId is validated by this point
         try {
             return jdbcTemplate.queryForObject(findIdByCodeAndClaimSql,
                     new MapSqlParameterSource()
                             .addValue("code", code)
                             .addValue("claimCode", claimCode), Integer.class);
-        } catch (EmptyResultDataAccessException ignore) {
+        } catch (final EmptyResultDataAccessException ignore) {
             return null;
         }
     }
